moves CustomCheckbox into Item.styles.js file
